Refactor: Rename and move app folder to view
